Switch to JAX L-BFGS optimizer via optax (34x faster than scipy float64)
Benchmarks on DAP044 (520 trials):
  JAX L-BFGS  float64:  -717 log_evd  8.8s   (recommended default)
  scipy NCG   float32:  -718 log_evd  34.9s
  scipy NCG   float64:  -699 log_evd  303s

- Default precision changed to float64 (matches scipy quality, much faster)
- optax>=0.2 added to requirements.txt
- App precision default updated to float64

+++ fit_all.py
@@ -79,7 +79,7 @@ def main():
                         help='Skip mice whose fit file already exists')
     parser.add_argument('--push', action='store_true',
                         help='git commit + push each fit as it completes')
-    parser.add_argument('--precision', default='float32',
+    parser.add_argument('--precision', default='float64',
                         choices=['float32', 'float64'],
                         help='JAX precision (default: float32 for speed)')
     parser.add_argument('--optimizer', default='jax',
